+++ .idea/misc.xml
@@ -13,6 +13,7 @@
         <entry key="..\:/Project/MobileProject/Prm391_Group_Project/app/src/main/res/layout/layout_header_navigation.xml" value="0.1" />
         <entry key="..\:/Project/MobileProject/Prm391_Group_Project/app/src/main/res/menu/home_menu.xml" value="0.264" />
         <entry key="..\:/Project/MobileProject/Prm391_Group_Project/app/src/main/res/menu/menu_navigation.xml" value="0.24895833333333334" />
+        <entry key="..\:/Project/MobileProject/Prm391_Group_Project/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1025" />
         <entry key="..\:/github/Prm391_Group_Project/app/src/main/res/layout/activity_main.xml" value="0.33" />
       </map>
     </option>
