+++ app/src/main/java/com/fpt/prm391groupproject/ProductAdapter.java
@@ -11,8 +11,12 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.fpt.prm391groupproject.fragment.ProfileFragment;
 import com.fpt.prm391groupproject.model.Product;
 
 import java.util.List;
@@ -20,15 +24,17 @@ import java.util.List;
 public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ProductViewHolder>{
     private List<Product> products;
     private Context context;
+    private FragmentActivity fragmentActivity;
 
-    public ProductAdapter(List<Product> products, Context context) {
+    public ProductAdapter(List<Product> products, FragmentActivity fragmentActivity) {
         this.products = products;
-        this.context = context;
+        this.fragmentActivity = fragmentActivity;
     }
 
     @NonNull
     @Override
     public ProductViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        context = parent.getContext();
         View v = LayoutInflater.from(context).inflate(R.layout.item_product, parent, false);
         return new ProductViewHolder(v);
     }
@@ -58,6 +64,20 @@ public class ProductAdapter extends RecyclerView.Adapter<ProductAdapter.ProductV
             tv_name = itemView.findViewById(R.id.tv_product_name);
             tv_price = itemView.findViewById(R.id.tv_product_price);
             btn_add =  itemView.findViewById(R.id.btn_add_to_cart);
+            btn_add.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View view) {
+
+                    String id =products.get(getAdapterPosition()).getId();
+
+                    Toast.makeText(view.getContext(),id,Toast.LENGTH_SHORT).show();
+
+                    FragmentTransaction transaction = fragmentActivity.getSupportFragmentManager().beginTransaction();
+                    transaction.replace(R.id.content_frame,new ProfileFragment(id));
+                    transaction.commit();
+
+                }
+            });
 
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
