+++ app/src/main/java/com/fpt/prm391groupproject/fragment/ProfileFragment.java
@@ -4,6 +4,7 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -12,9 +13,19 @@ import androidx.fragment.app.Fragment;
 import com.fpt.prm391groupproject.R;
 
 public class ProfileFragment extends Fragment {
+    TextView textView;
+    private String id;
+
+    public ProfileFragment(String id) {
+        this.id = id;
+    }
+
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_profile,container,false);
+        View view = inflater.inflate(R.layout.fragment_profile,container,false);
+        textView = view.findViewById(R.id.test_text);
+        textView.setText(id);
+        return view;
     }
 }
