+++ app/src/main/res/layout/fragment_home.xml
@@ -10,6 +10,6 @@
         android:id="@+id/recycle_view_home"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="1000dp"
+        android:layout_marginTop="50dp"
         />
 </LinearLayout>
