+++ app/src/main/res/layout/fragment_profile.xml
@@ -7,9 +7,10 @@
     android:background="@color/green">
 
     <TextView
+        android:id="@+id/test_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textColor="@color/white"
         android:textSize="20sp"
-        android:text="Profile fragment"/>
+        android:text="Hello ae"/>
 </LinearLayout>
